made bottom and top header into components and modular
